More block cleanup
For the most part I’m only using a couple of templates.  There are one or two minor customizations that are only used for one or two newsletters so it’s not worth putting those in components (right now).  Like the header of Food/Beverage and the weird slanted border on IEN Update & Weekly