Update chunk generation & add call for leftover adds
Update chunk logic and how it calculates how big the chunk should be…

$ const chunkNumber = (banners.length) ? ((nodes.length % banners.length === 0) ? nodes.length / banners.length - 1 : Math.floor(nodes.length / banners.length)) : nodes.length;

 - if no banners.length return nodes.length to render all content in one chunk
 - if banners && length evenly devids into nodes subtract one to ensure last ad has content below
 - else round down nodes.length/banners.length

Yes this can make the chunkNumber 0, but the next line accounts for that and sets minimum chunk to 1.

All said and done also account for when there are more banners than chunks available and loop over and render extra ads.